Separated keypress.c and utils.c
